# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: odo API
  version: 0.1.0
  summary: API-first library access middleware/proxy MVP
  description: Management and configuration for odo happen through documented JSON APIs. The admin UI uses these same APIs.
servers:
  - url: http://127.0.0.1:8080
security: []
paths:
  /api/v1/health:
    get:
      summary: Health check
      operationId: getHealth
      security: []
      responses:
        "200":
          description: Service health
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"
  /api/v1/resources:
    get:
      summary: List resources
      operationId: listResources
      responses:
        "200":
          description: Resource list
          content:
            application/json:
              schema:
                type: object
                required: [resources]
                properties:
                  resources:
                    type: array
                    items:
                      $ref: "#/components/schemas/Resource"
    post:
      summary: Create or update a resource
      operationId: upsertResource
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Resource"
      responses:
        "200":
          description: Upserted resource
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resource"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
  /api/v1/resources/{id}:
    get:
      summary: Get one resource
      operationId: getResource
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Resource
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resource"
        "404":
          $ref: "#/components/responses/Error"
    put:
      summary: Create or update a resource at the given ID
      operationId: putResource
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Resource"
      responses:
        "200":
          description: Upserted resource
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resource"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
    delete:
      summary: Delete a resource
      operationId: deleteResource
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Resource deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResourceResponse"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
  /api/v1/config/validate:
    post:
      summary: Validate dropped resource config files without writing to SQLite
      operationId: validateConfig
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Config validation results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConfigValidationResponse"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
  /api/v1/config/import:
    post:
      summary: Import dropped resource config files
      operationId: importConfig
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Config import results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConfigImportResponse"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
  /api/v1/config/revisions:
    get:
      summary: List recent config revisions
      operationId: listConfigRevisions
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Recent config revisions
          content:
            application/json:
              schema:
                type: object
                required: [revisions]
                properties:
                  revisions:
                    type: array
                    items:
                      $ref: "#/components/schemas/ConfigRevisionSummary"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
  /api/v1/config/revisions/{id}:
    get:
      summary: Get config revision detail
      operationId: getConfigRevision
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
      responses:
        "200":
          description: Config revision detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConfigRevisionDetail"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
  /api/v1/rules/test-url:
    post:
      summary: Test whether a URL matches an active resource rule
      operationId: testURL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TestURLRequest"
      responses:
        "200":
          description: URL match result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchResult"
        "400":
          $ref: "#/components/responses/Error"
  /api/v1/proxy/test-fetch:
    post:
      summary: Fetch a bounded proxy troubleshooting preview
      operationId: proxyTestFetch
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProxyTestFetchRequest"
      responses:
        "200":
          description: Proxy fetch preview or safe denial result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProxyTestFetchResponse"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "502":
          $ref: "#/components/responses/Error"
  /api/v1/logs/access/recent:
    get:
      summary: List recent privacy-filtered access log entries
      operationId: recentAccessLogs
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Recent in-memory access log entries
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecentAccessLogsResponse"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
  /api/v1/diagnostics/proxy/recent:
    get:
      summary: List recent proxy diagnostics
      operationId: recentProxyDiagnostics
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Recent proxy diagnostics, currently empty in the MVP
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecentProxyDiagnosticsResponse"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
  /p:
    get:
      summary: Minimal proxy fetch
      description: Validates and allowlist-checks the target URL, performs a safe outbound GET, streams the upstream response, applies minimal HTML/CSS asset URL rewriting, and uses a server-side per-session cookie jar. Upstream cookies are not passed directly to the browser. JavaScript rewriting is not implemented yet.
      operationId: proxyFetch
      security: []
      parameters:
        - name: url
          in: query
          required: true
          schema:
            type: string
            format: uri
      responses:
        "200":
          description: Upstream response body and safe response headers
        "403":
          description: URL is not allowed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchResult"
        "405":
          $ref: "#/components/responses/Error"
        "502":
          $ref: "#/components/responses/Error"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API key
  responses:
    Error:
      description: Error response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    HealthResponse:
      type: object
      required: [status, time]
      properties:
        status:
          type: string
          examples: [ok]
        time:
          type: string
          format: date-time
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    DeleteResourceResponse:
      type: object
      required: [deleted, id]
      properties:
        deleted:
          type: boolean
        id:
          type: string
    Resource:
      type: object
      required: [id, name, status, domains]
      properties:
        id:
          type: string
          examples: [jstor]
        name:
          type: string
          examples: [JSTOR]
        status:
          type: string
          default: active
          examples: [active]
        description:
          type: string
        domains:
          type: array
          items:
            $ref: "#/components/schemas/DomainRule"
        sample_urls:
          type: array
          items:
            type: string
            format: uri
    DomainRule:
      type: object
      required: [host, match, role, action]
      properties:
        host:
          type: string
          examples: [www.jstor.org]
        match:
          type: string
          enum: [exact, subdomain]
          default: exact
        role:
          type: string
          enum: [content, asset, api, auth, redirect, external, blocked]
          default: content
          examples: [content]
        action:
          type: string
          enum: [proxy, allow, block]
          default: proxy
          examples: [proxy]
        reason:
          type: string
    ConfigValidationResponse:
      type: object
      required: [valid, config_dir, results]
      properties:
        valid:
          type: boolean
        config_dir:
          type: string
        results:
          type: array
          items:
            $ref: "#/components/schemas/ConfigValidationResult"
    ConfigValidationResult:
      type: object
      required: [file, valid, resource_id, errors]
      properties:
        file:
          type: string
        valid:
          type: boolean
        resource_id:
          type: string
        errors:
          type: array
          items:
            type: string
    ConfigImportResponse:
      type: object
      required: [results]
      properties:
        results:
          type: array
          items:
            $ref: "#/components/schemas/ConfigImportResult"
    ConfigImportResult:
      type: object
      required: [file, imported]
      properties:
        file:
          type: string
        resource_id:
          type: string
        imported:
          type: boolean
        error:
          type: string
    ConfigRevisionSummary:
      type: object
      required: [id, ts, source, status, summary]
      properties:
        id:
          type: integer
        ts:
          type: string
          format: date-time
        source:
          type: string
          examples: [file-import]
        status:
          type: string
          examples: [applied]
        summary:
          type: string
          examples: [imported=1 errors=0]
    ConfigRevisionDetail:
      allOf:
        - $ref: "#/components/schemas/ConfigRevisionSummary"
        - type: object
          required: [config_json]
          properties:
            config_json:
              type: string
              description: JSON string containing normalized imported resource configs.
    TestURLRequest:
      type: object
      required: [url]
      properties:
        url:
          type: string
          format: uri
          examples: [https://www.jstor.org/stable/example]
    ProxyTestFetchRequest:
      type: object
      required: [url]
      properties:
        url:
          type: string
          format: uri
          examples: [https://www.jstor.org/]
    ProxyTestFetchResponse:
      type: object
      required: [allowed]
      properties:
        allowed:
          type: boolean
        error:
          type: string
        reason:
          type: string
        status:
          type: integer
        target_host:
          type: string
        resource_id:
          type: string
        content_type:
          type: string
        body_preview:
          type: string
        body_preview_truncated:
          type: boolean
        headers:
          type: object
          additionalProperties:
            type: string
    RecentAccessLogsResponse:
      type: object
      required: [entries]
      properties:
        entries:
          type: array
          items:
            $ref: "#/components/schemas/AccessLogEntry"
    RecentProxyDiagnosticsResponse:
      type: object
      required: [entries]
      properties:
        entries:
          type: array
          items:
            type: object
            additionalProperties: true
    AccessLogEntry:
      type: object
      required: [ts, request_id, remote_ip, method, route, status, bytes, duration_ms]
      properties:
        ts:
          type: string
          format: date-time
        request_id:
          type: string
        remote_ip:
          type: string
        method:
          type: string
        route:
          type: string
        status:
          type: integer
        bytes:
          type: integer
        duration_ms:
          type: integer
        target_host:
          type: string
        resource_id:
          type: string
        rule_host:
          type: string
        rule_match:
          type: string
        decision:
          type: string
        denial_reason:
          type: string
        upstream_status:
          type: integer
    MatchResult:
      type: object
      required: [allowed, reason]
      properties:
        allowed:
          type: boolean
        blocked:
          type: boolean
        host:
          type: string
        resource_id:
          type: string
        rule_host:
          type: string
        rule_match:
          type: string
          enum: [exact, subdomain]
        role:
          type: string
          enum: [content, asset, api, auth, redirect, external, blocked]
        action:
          type: string
          enum: [proxy, allow, block]
        matched_rule:
          $ref: "#/components/schemas/DomainRule"
        reason:
          type: string